Fixed KDE issues now
