+++ pbjam/__init__.py
@@ -4,8 +4,8 @@
 import os
 PACKAGEDIR = os.path.abspath(os.path.dirname(__file__))
 
-from .guess_epsilon import epsilon
-from .jar import session
+from .kde import kde
+from .session import session
 from .asy_peakbag import asymp_spec_model
 from .peakbag import peakbag
 from .star import star
